Rework fswatcher.WatchOn(...)
